+++ app/.eslintrc.cjs
@@ -1,17 +1,26 @@
 module.exports = {
   root: true,
-  extends: [
-    'react-app',
-    'react-app/jest',
-    'airbnb',
-    'airbnb/hooks',
-    'plugin:prettier/recommended',
-  ],
+  env: {
+    browser: true,
+    es2021: true,
+    node: true,
+    jest: true,
+  },
+  extends: ['airbnb', 'airbnb/hooks', 'plugin:prettier/recommended'],
+  parserOptions: {
+    ecmaVersion: 'latest',
+    sourceType: 'module',
+    ecmaFeatures: {
+      jsx: true,
+    },
+  },
+  settings: {
+    react: {
+      version: 'detect',
+    },
+  },
   rules: {
-    'react/jsx-filename-extension': [
-      'error',
-      { extensions: ['.jsx', '.js'] },
-    ],
+    'react/jsx-filename-extension': ['error', { extensions: ['.jsx', '.js'] }],
     'react/react-in-jsx-scope': 'off',
     'react/jsx-one-expression-per-line': 'off',
     'react/no-unstable-nested-components': ['warn', { allowAsProps: true }],
@@ -24,6 +33,7 @@ module.exports = {
       {
         devDependencies: [
           '**/setupTests.js',
+          '**/vite.config.js',
           '**/tailwind.config.js',
           '**/postcss.config.js',
           '**/*.test.js',